move most of code from format.go --> gopher.go
